SAR Validation: Condition keys, Resource patterns, Resource targets (#29)
### Overview

This PR makes fairly broad changes across the codebase in order to more
accurately enforce the per-service authorization rules of the AWS SAR.

A summary of the changes include:
* Rewrite `sar.py` into `sar_v2.py`, which uses the
incomplete-but-more-accurate [service
reference](https://docs.aws.amazon.com/en_us/service-authorization/latest/reference/service-reference.html)
* Avoid over-processing of the SAR data when loading into `yams`; keep
it roughly as-is and use a separate `map`-based index
* During simulation, validate condition keys against the `Action`
represented in the `AuthContext`. Even if particular keys are present in
the `AuthContext` of a request, they will not be resolved unless that
particular API call supports them
* Valid condition keys can also be based on a particular `Resource` type
being used for evaluation, so for multi-resource API calls we need to
match the provided `Resource` field against the types respected by the
API call
* During simulation, perform further validation of the `AuthContext`
itself (e.g. are all the required fields present? Was a `Resource`
provided for an API call that is not targeted at a `Resource`? Did the
wrong `Resource` type get presented for a particular API call? etc
* Adding significantly more tests for default-behavior condition key
retrieval
* Use a case-insensitive `PropertyBag` to more accurately implement the
case-insensitivity of `Condition` keys